requirements.txt
# Azure AI Foundry (Foundry Agent Service)
azure-identity>=1.17.1,<2.0
azure-ai-projects==2.0.0b4
azure-ai-agents==1.2.0b5
python-dotenv>=1.0.1,<2.0
azure-identity-broker>=1.0.0,<2.0
openai